+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/virtual-users/branches/compare/master%0Drelease-2.1.5
 
+### Fixed
+- Print relative paths for dumps in WebDriverDiagnostics as a workaround for [JPERF-158].
+
+[JPERF-158]: https://ecosystem.atlassian.net/browse/JPERF-158
+
 ## [2.1.5] - 2018-10-25
 [2.1.5]: https://bitbucket.org/atlassian/virtual-users/branches/compare/release-2.1.5%0Drelease-2.1.4